adjust the js files structure
